+++ src/main/java/spofo/portfolio/service/PortfolioServiceImpl.java
@@ -7,6 +7,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import spofo.global.domain.exception.PortfolioNotFound;
+import spofo.portfolio.controller.port.PortfolioService;
 import spofo.portfolio.domain.Portfolio;
 import spofo.portfolio.domain.PortfolioCreate;
 import spofo.portfolio.domain.PortfolioStatistic;
