+++ src/main/java/spofo/tradelog/controller/port/TradeLogService.java
@@ -1,4 +1,4 @@
-package spofo.tradelog.service;
+package spofo.tradelog.controller.port;
 
 import java.util.List;
 import spofo.tradelog.controller.response.TradeLogResponse;
@@ -7,6 +7,6 @@ import spofo.tradelog.domain.CreateTradeLogRequest;
 public interface TradeLogService {
 
     void createTradeLog(CreateTradeLogRequest createTradeLogRequest);
-    
+
     List<TradeLogResponse> getTradeLogs(Long stockId);
 }
